refactor(sql/tables): split tables across multiple files
