+++ sql/tables/__init__.py
@@ -1,0 +1,6 @@
+from .distances import Distances
+from .jobs import Jobs
+from .keywords import Keywords
+from .metadata import Metadata
+from .time_stamps import TimeStamps
+from .base_table import BaseTableForJobScrapper
